Annual and cumulative percent change

Summarize annual and cumulative percent change in risk

apc(x)

# S3 method for surveil
apc(x)

# S3 method for stand_surveil
apc(x)

# S3 method for apc
print(x, digits = 1, max = 10, ...)

# S3 method for apc
plot(
  x,
  cumulative = FALSE,
  style = c("mean_qi", "lines"),
  M = 250,
  col = "black",
  fill = "black",
  alpha,
  lwd,
  base_size = 14,
  ...
)

Arguments

x
An apc object returned by apc

digits
Print this many digits (passed to print.data.frame)

max
Maximum number of time periods (rows) to print

...
Additional arguments

cumulative
Plot cumulative percent change? Defaults to cumulative = FALSE

style
If style = "mean_qi", then the posterior mean and 95 percent credible interval will be plotted; if style = "lines", then M samples from the joint probability distribution will be plotted.

M
If style = "lines", then M is the number of samples from the posterior distribution that will be plotted; the default is M = 250.

col
Line color

fill
Fill color for the 95 percent credible interval

alpha
For style = "mean_qi", this controls the transparency for the credible interval (passed to geom_ribbon) and defaults to alpha = 0.5; for style = "lines", this controls the transparency of the lines and defaults to alpha = 0.7.

lwd
Line width

base_size
Size of plot attributes, passed to `theme_classic

Value

An apc (list) object containing the following data frames:

apc

A data frame containing a summary of the posterior distribution for period-specific percent change. This contains the posterior mean (apc) 95 percent credible intervals (lwr and upr bounds).

cpc

A data frame containing a summary of the posterior distribution for the cumulative percent change in risk at each time period. This contains the posterior mean (cpc) and 95 percent credible interval (lwr and upr bounds).

apc_samples

MCMC samples from the posterior distribution for period percent change

cpc_samples

MCMC samples from the posterior distribution for cumulative percent change

plot.apc

The plot method returns a ggplot

Details

print.apc

Any additional arguments (...) to the print method will be passed to print.data.frame

plot.apc

Any additional arguments (...) will be passed to theme.

See also

Examples

data(cancer)
# \donttest{
 fit <- stan_rw(cancer, time = Year, group = Age, iter = 2e3)
 x <- apc(fit)
 print(x)
 plot(x, cumulative = TRUE)
# }