Time series models for mortality and disease incidence

Model time-varying disease risk given time series of case (or death) counts and population at risk.

stan_rw(
  data,
  group,
  time,
  cor = FALSE,
  prior = list(),
  chains = 4,
  cores = 1,
  iter = 3000,
  refresh = 1500,
  control = list(adapt_delta = 0.975),
  ...
)

Arguments

data
A data.frame containing the following columns:

Count

Number of cases or deaths; this column must be named 'Count'.

Population

Size of population at risk; this column must be named 'Population'.

time

Time period indicator. (Provide the (unquoted) column name using the time argument.)

group

Optional grouping variable. (Provide the (unquoted) column name using the group argument.)

group
If data is aggregated by demographic group, provide the (unquoted) name of the column in data containing the grouping structure, such as age brackets or race-ethnicity. E.g., if data has column names Year, Race, Cases, and Population, then you would provide group = Race.

time
Specify the (unquoted) name of the time variable in data, as in time = Year. This variable must be numeric-alike (i.e., as.numeric(data$time) will not fail).

cor
For correlated random walks use cor = TRUE; default value is FALSE. Note this only applies when the group argument is used.

prior
Optionally provide a named list with prior parameters. If any of the following items are missing, default priors will be assigned and printed to the console.

eta_1

The first value of log-risk in each series must be assigned a Gaussian prior probability distribution. Provide the location and scale parameters for each demographic group in a list, where each parameter is a k-length vector.For example, with k=2 demographic groups, the following code will assign priors of normal(-5, 5) to the starting values of both series: prior = list(eta_1 = normal(location = -6.5, scale = 5, k = 2). Note, eta is the log-rate, so centering the prior for eta_1 on -5 is equivalent to centering the prior rate on exp(-6.5)*100,000 = 150 cases per 100,000 person-years at risk.

sigma

Each demographic group has a scale parameter assigned to its log-rate. This is the scale of the annual deviations from the previous year's log-rate. The scale parameters are assigned independent half-normal prior distributions (these half normal distributions are restricted to be positive-valued only).

omega

If cor = TRUE, an LKJ prior is assigned to the correlation matrix, Omega.

chains
Number of independent MCMC chains to initiate (passed to sampling).

cores
The number of cores to use when executing the Markov chains in parallel (passed to sampling).

iter
Total number of MCMC iterations. Warmup draws are automatically half of iter.

refresh
How often to print the MCMC sampling progress to the console.

control
A named list of parameters to control Stan's sampling behavior. The most common parameters to control are adapt_delta, which may be raised to address divergent transitions, and max_treedepth. For example, control = list(adapt_delta = 0.99, max_treedepth = 13), may be a reasonable specification to address a divergent transitions or maximum treedepth warning from Stan.

...
Other arguments passed to sampling.

Source

Brandt P, Williams JT. Multiple time series models. Thousand Oaks, CA: SAGE Publications, 2007.

Clayton, DG. Generalized linear mixed models. In: Gilks WR, Richardson S, Spiegelhalter DJ, editors. Markov Chain Monte Carlo in Practice: Interdisciplinary Statistics. Boca Raton, FL: CRC Press, 1996. p. 275-302.

Stan Development Team. Stan Modeling Language Users Guide and Reference Manual, 2.28. 2021. https://mc-stan.org

Details

For time t = 1,...n, the models have Poisson likelihoods for the case counts, given log-risk eta and population at tirks P; the log-risk is modeled using the first-difference prior (a.k.a. a random-walk).

       y_t ~ Poisson(p_t * exp(eta_t))
       eta_t ~ Normal(eta_{t-1}, sigma)
       eta_1 ~ Normal(-5, 10)
       sigma ~ Normal(0, 1) (0, Inf]

This style of model has been discussed in Bayesian (bio)statistics for quite some time. See Clayton (1996).

The above model can be used for multiple distinct groups; in that case, each group will have its own independent time series model. However, it is also possible to add a correlation structure to that set of models. Let Y_t be a k-length vector of observations for each of k groups at time t (the capital letter now indicates a vector), then:

       Y_t ~ Poisson(P_t * exp(Eta_t))
       Eta_t ~ MVNormal(Eta_{t-1}, Sigma)
       Eta_1 ~ Normal(-5, 10)
       Sigma = diag(sigma) Omega diag(sigma)
       Omega ~ LKJ(2)
       sigma ~ Normal(0, 1) (0, Inf)

where Omega is a correlation matrix and diag(sigma) is a diagonal matrix with scale parameters on the diagonal. This was adopted from Brandt and Williams (2007); for the LKJ prior, see the Stan Users Guide and Reference Manual.

Author

Connor Donegan (Connor.Donegan@UTSouthwestern.edu)

Examples

if (FALSE) {
data(msa)
dfw <- msa[grep("Dallas", msa$MSA), ]
fit <- stan_rw(dfw, time = Year, group = Race)

## Dont use results if you see a warning for divergent transitions;
## try to fix the problem by adjusting adapt_delta:
fit <- stan_rw(dfw, time = Year, group = Race,
               control = list(adapt_delta = 0.99))

## print the underlying Stan model, with MCMC diagnostics
print(fit$samples)

## print the surveil object
print(fit)
head(fit$summary)

## plot time trends
plot(fit)

## Inequality analysis
Ti <- theil(fit)
plot(Ti)
print(Ti)

gd <- group_diff(fit, target = "Black or African American", reference = "White")
plot(gd)
print(gd, scale = 10e3)

## age-specific rates and cumulative percent change
data(cancer)
fit <- stan_rw(cancer,
               time = Year,
               group = Age
               )
fit_apc <- apc(fit)
plot(fit_apc, cumulative = TRUE)

# age-standardized rates
data(standard)
fit_stands <- standardize(fit, label = standard$age, standard_pop = standard$standard_pop)
print(fit_stands)
plot(fit_stands)
fit_stands_apc <- apc(fit_stands)
plot(fit_stands_apc)
}