Standardized rates

Convert surveil model results to age standardized rates using a fixed age distribution

standardize(x, label, standard_pop)

Arguments

x
A fitted surveil model

label
Labels (character strings) for the age groups that correspond to the values of stand_pop. The labels must match the grouping variable used to fit the model (i.e., all(label %in% names(x$data$cases)) must be true).

standard_pop
Standard population values corresponding to the age groups specified by label

Value

A list, also of class "stand_surveil", containing a summary data frame (mean and 95 percent credible intervals) (named summary), a data frame of MCMC samples of standardized rates per time period (named samples), and the user-provided labels and standard population sizes (named label and standard_pop). In addition, all of the items from the user-provided surveil model are automatically appended to the list.

See also

Examples

data(cancer)
data(standard)

head(standard)
#>   age.id   age standard_pop
#> 1      1    <1        13818
#> 2      2   1-4        55317
#> 3      3   5-9        72533
#> 4      4 10-14        73032
#> 5      5 15-19        72169
#> 6      6 20-24        66478
head(cancer)
#>   Year   Age Count Population
#> 1 1999    <1   866    3708753
#> 2 1999   1-4  2959   14991152
#> 3 1999   5-9  2226   20146188
#> 4 1999 10-14  2447   19742631
#> 5 1999 15-19  3875   19585857
#> 6 1999 20-24  5969   18148795

if (FALSE) {
fit <- stan_rw(cancer,
              time = Year,
              group = Age
              )

stands <- standardize(fit,
                      label = standard$age,
                      standard_pop = standard$standard_pop)
print(stands)
plot(stands)
}